[10.x] Update test for Laravel 9 (#1509)
* Update test for Laravel 9

* wip

* wip
